feat: integrate error handling service with ui
